+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 * Increment `@folio/stripes-cli` to `v2`. Refs UIU-2047.
 * Inconsistent behavior of payment action when transaction info is blank. Refs UIU-2048.
 * Fix wrong value on reset data for Refunds to process manually report (CSV). Refs UIU-2050.
+* Add Callout for Refund report. Refs UIU-2035.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.0...v5.0.1)